Moved ClangViewParse to source_clang.*.
