+++ src/notebook.h
@@ -4,7 +4,7 @@
 #include <iostream>
 #include "gtkmm.h"
 #include "source.h"
-#include <boost/algorithm/string/case_conv.hpp>
+#include "source_clang.h"
 #include <type_traits>
 #include <map>
 #include <sigc++/sigc++.h>
