+++ src/source.h
@@ -1,20 +1,14 @@
 #ifndef JUCI_SOURCE_H_
 #define JUCI_SOURCE_H_
-#include <iostream>
 #include <unordered_map>
 #include <vector>
 #include "gtkmm.h"
 #include "clangmm.h"
-#include <thread>
-#include <mutex>
 #include <string>
-#include <atomic>
 #include "gtksourceviewmm.h"
 #include "terminal.h"
 #include "tooltips.h"
 #include "selectiondialog.h"
-#include <set>
-#include <regex>
 #include <aspell.h>
 #include <boost/property_tree/xml_parser.hpp>
 
@@ -200,119 +194,5 @@ namespace Source {
     
     void parse_language_file(Glib::RefPtr<CompletionBuffer> &completion_buffer, bool &has_context_class, const boost::property_tree::ptree &pt);
   };
-  
-  class ClangViewParse : public View {
-  public:
-    class TokenRange {
-    public:
-      TokenRange(std::pair<clang::Offset, clang::Offset> offsets, int kind):
-        offsets(offsets), kind(kind) {}
-      std::pair<clang::Offset, clang::Offset> offsets;
-      int kind;
-    };
-    
-    ClangViewParse(const boost::filesystem::path &file_path, const boost::filesystem::path& project_path, Glib::RefPtr<Gsv::Language> language);
-    ~ClangViewParse();
-    void configure();
-    
-    void start_reparse();
-    bool reparse_needed=false;
-  protected:
-    void init_parse();
-    bool on_key_press_event(GdkEventKey* key);
-    std::unique_ptr<clang::TranslationUnit> clang_tu;
-    std::mutex parsing_mutex;
-    std::unique_ptr<clang::Tokens> clang_tokens;
-    sigc::connection delayed_reparse_connection;
-    
-    std::shared_ptr<Terminal::InProgress> parsing_in_progress;
-    std::thread parse_thread;
-    std::atomic<bool> parse_thread_stop;
-    std::atomic<bool> parse_error;
-    
-    void show_diagnostic_tooltips(const Gdk::Rectangle &rectangle);
-    void show_type_tooltips(const Gdk::Rectangle &rectangle);
-    
-    std::regex bracket_regex;
-    std::regex no_bracket_statement_regex;
-    std::regex no_bracket_no_para_statement_regex;
-    
-    std::set<int> diagnostic_offsets;
-    std::vector<FixIt> fix_its;
-  private:
-    std::map<std::string, std::string> get_buffer_map() const;
-    void update_syntax();
-    std::set<std::string> last_syntax_tags;
-    void update_diagnostics();
-    
-    static clang::Index clang_index;
-    std::vector<std::string> get_compilation_commands();
-    
-    Glib::Dispatcher parse_done;
-    Glib::Dispatcher parse_start;
-    Glib::Dispatcher parse_fail;
-    std::map<std::string, std::string> parse_thread_buffer_map;
-    std::mutex parse_thread_buffer_map_mutex;
-    std::atomic<bool> parse_thread_go;
-    std::atomic<bool> parse_thread_mapped;
-  };
-    
-  class ClangViewAutocomplete : public ClangViewParse {
-  public:
-    class AutoCompleteData {
-    public:
-      explicit AutoCompleteData(const std::vector<clang::CompletionChunk> &chunks) :
-        chunks(chunks) { }
-      std::vector<clang::CompletionChunk> chunks;
-      std::string brief_comments;
-    };
-    
-    ClangViewAutocomplete(const boost::filesystem::path &file_path, const boost::filesystem::path& project_path, Glib::RefPtr<Gsv::Language> language);
-    void async_delete();
-    bool restart_parse();
-  protected:
-    bool on_key_press_event(GdkEventKey* key);
-    std::thread autocomplete_thread;
-  private:
-    void start_autocomplete();
-    void autocomplete();
-    std::unique_ptr<CompletionDialog> completion_dialog;
-    bool completion_dialog_shown=false;
-    std::vector<AutoCompleteData> get_autocomplete_suggestions(int line_number, int column, std::map<std::string, std::string>& buffer_map);
-    Glib::Dispatcher autocomplete_done;
-    sigc::connection autocomplete_done_connection;
-    Glib::Dispatcher autocomplete_fail;
-    sigc::connection autocomplete_fail_connection;
-    bool autocomplete_starting=false;
-    std::atomic<bool> autocomplete_cancel_starting;
-    guint last_keyval=0;
-    std::string prefix;
-    std::mutex prefix_mutex;
-    
-    Glib::Dispatcher do_delete_object;
-    Glib::Dispatcher do_restart_parse;
-    std::thread delete_thread;
-    std::thread restart_parse_thread;
-    bool restart_parse_running=false;
-  };
-
-  class ClangViewRefactor : public ClangViewAutocomplete {
-  public:
-    ClangViewRefactor(const boost::filesystem::path &file_path, const boost::filesystem::path& project_path, Glib::RefPtr<Gsv::Language> language);
-    ~ClangViewRefactor();
-  private:
-    std::list<std::pair<Glib::RefPtr<Gtk::TextMark>, Glib::RefPtr<Gtk::TextMark> > > similar_token_marks;
-    void tag_similar_tokens(const Token &token);
-    Glib::RefPtr<Gtk::TextTag> similar_tokens_tag;
-    Token last_tagged_token;
-    sigc::connection delayed_tag_similar_tokens_connection;
-    std::unique_ptr<SelectionDialog> selection_dialog;
-    bool renaming=false;
-  };
-  
-  class ClangView : public ClangViewRefactor {
-  public:
-    ClangView(const boost::filesystem::path &file_path, const boost::filesystem::path& project_path, Glib::RefPtr<Gsv::Language> language);
-  };
-};  // class Source
+}
 #endif  // JUCI_SOURCE_H_
